refactor(behavior_path_planner): move createObjectDebug to safety check
