+++ planning/behavior_path_planner/src/marker_utils/utils.cpp
@@ -42,15 +42,6 @@ using tier4_autoware_utils::createMarkerScale;
 using tier4_autoware_utils::createPoint;
 using visualization_msgs::msg::Marker;
 
-CollisionCheckDebugPair createObjectDebug(const ExtendedPredictedObject & obj)
-{
-  CollisionCheckDebug debug;
-  debug.current_obj_pose = obj.initial_pose.pose;
-  debug.extended_obj_polygon = tier4_autoware_utils::toPolygon2d(obj.initial_pose.pose, obj.shape);
-  debug.current_twist = obj.initial_twist.twist;
-  return {tier4_autoware_utils::toHexString(obj.uuid), debug};
-}
-
 void updateCollisionCheckDebugMap(
   CollisionCheckDebugMap & debug_map, CollisionCheckDebugPair & object_debug, bool is_safe)
 {
