+++ planning/behavior_path_planner/src/utils/path_safety_checker/safety_check.cpp
@@ -17,6 +17,7 @@
 #include "interpolation/linear_interpolation.hpp"
 #include "motion_utils/trajectory/trajectory.hpp"
 #include "tier4_autoware_utils/geometry/boost_polygon_utils.hpp"
+#include "tier4_autoware_utils/ros/uuid_helper.hpp"
 
 #include <boost/geometry/algorithms/intersects.hpp>
 #include <boost/geometry/algorithms/overlaps.hpp>
@@ -392,4 +393,14 @@ bool checkPolygonsIntersects(
   }
   return false;
 }
+
+CollisionCheckDebugPair createObjectDebug(const ExtendedPredictedObject & obj)
+{
+  CollisionCheckDebug debug;
+  debug.current_obj_pose = obj.initial_pose.pose;
+  debug.extended_obj_polygon = tier4_autoware_utils::toPolygon2d(obj.initial_pose.pose, obj.shape);
+  debug.current_twist = obj.initial_twist.twist;
+  return {tier4_autoware_utils::toHexString(obj.uuid), debug};
+}
+
 }  // namespace behavior_path_planner::utils::path_safety_checker
